Refactored name to DelayNotificationService
